Move deployment package to ../infrastructure package
