+++ src/main/groovy/com/cloudogu/gitops/infrastructure/deployment/DeploymentStrategy.groovy
@@ -1,4 +1,4 @@
-package com.cloudogu.gitops.features.deployment
+package com.cloudogu.gitops.infrastructure.deployment
 
 import java.nio.file.Path
