+++ src/test/groovy/com/cloudogu/gitops/ApplicationConfiguratorTest.groovy
@@ -1,29 +1,30 @@
 package com.cloudogu.gitops
 
+import static com.github.stefanbirkner.systemlambda.SystemLambda.withEnvironmentVariable
+import static groovy.test.GroovyAssert.shouldFail
+import static org.assertj.core.api.Assertions.assertThat
+
 import com.cloudogu.gitops.config.ApplicationConfigurator
 import com.cloudogu.gitops.config.CommonFeatureConfig
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.features.ContentLoader
 import com.cloudogu.gitops.features.Jenkins
 import com.cloudogu.gitops.features.argocd.ArgoCD
-import com.cloudogu.gitops.features.deployment.DeploymentStrategy
 import com.cloudogu.gitops.features.git.GitHandler
 import com.cloudogu.gitops.features.git.config.ScmTenantSchema
 import com.cloudogu.gitops.git.GitRepoFactory
-import com.cloudogu.gitops.utils.FileSystemUtils
+import com.cloudogu.gitops.infrastructure.deployment.DeploymentStrategy
 import com.cloudogu.gitops.kubernetes.api.HelmClient
 import com.cloudogu.gitops.kubernetes.api.K8sClient
+import com.cloudogu.gitops.utils.FileSystemUtils
 import com.cloudogu.gitops.utils.TestLogger
 import com.cloudogu.gitops.utils.git.GitHandlerForTests
 import com.cloudogu.gitops.utils.git.ScmManagerMock
+
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.mockito.Mock
 import org.mockito.Mockito
-
-import static com.github.stefanbirkner.systemlambda.SystemLambda.withEnvironmentVariable
-import static groovy.test.GroovyAssert.shouldFail
-import static org.assertj.core.api.Assertions.assertThat
 
 class ApplicationConfiguratorTest {
