+++ src/test/groovy/com/cloudogu/gitops/features/deployment/DeployerTest.groovy
@@ -5,6 +5,10 @@ import static org.mockito.ArgumentMatchers.anyString
 import static org.mockito.Mockito.*
 
 import com.cloudogu.gitops.config.Config
+import com.cloudogu.gitops.infrastructure.deployment.ArgoCdApplicationStrategy
+import com.cloudogu.gitops.infrastructure.deployment.Deployer
+import com.cloudogu.gitops.infrastructure.deployment.DeploymentStrategy
+import com.cloudogu.gitops.infrastructure.deployment.HelmStrategy
 
 import java.nio.file.Path
