+++ src/test/groovy/com/cloudogu/gitops/features/deployment/HelmStrategyTest.groovy
@@ -1,17 +1,19 @@
 package com.cloudogu.gitops.features.deployment
 
-import org.junit.jupiter.api.Test
+import static groovy.test.GroovyAssert.shouldFail
+import static org.assertj.core.api.Assertions.assertThat
+import static org.mockito.Mockito.mock
+import static org.mockito.Mockito.verify
 
 import com.cloudogu.gitops.config.Config
+import com.cloudogu.gitops.infrastructure.deployment.DeploymentStrategy
+import com.cloudogu.gitops.infrastructure.deployment.HelmStrategy
 import com.cloudogu.gitops.kubernetes.api.HelmClient
 
 import java.nio.file.Files
 import java.nio.file.Path
 
-import static groovy.test.GroovyAssert.shouldFail
-import static org.assertj.core.api.Assertions.assertThat
-import static org.mockito.Mockito.mock
-import static org.mockito.Mockito.verify
+import org.junit.jupiter.api.Test
 
 class HelmStrategyTest {
