+++ src/test/groovy/com/cloudogu/gitops/git/providers/scmmanager/ScmManagerSetupTest.groovy
@@ -1,18 +1,18 @@
 package com.cloudogu.gitops.git.providers.scmmanager
 
-import org.junit.jupiter.api.Test
-import retrofit2.Call
-import retrofit2.Response
-
-import com.cloudogu.gitops.config.Config
-import com.cloudogu.gitops.features.deployment.HelmStrategy
-import com.cloudogu.gitops.git.providers.scmmanager.api.PluginApi
-import com.cloudogu.gitops.git.providers.scmmanager.api.ScmManagerApi
-import com.cloudogu.gitops.git.providers.scmmanager.api.ScmManagerApiClient
-
 import static org.mockito.ArgumentMatchers.any
 import static org.mockito.ArgumentMatchers.eq
 import static org.mockito.Mockito.*
+
+import com.cloudogu.gitops.config.Config
+import com.cloudogu.gitops.git.providers.scmmanager.api.PluginApi
+import com.cloudogu.gitops.git.providers.scmmanager.api.ScmManagerApi
+import com.cloudogu.gitops.git.providers.scmmanager.api.ScmManagerApiClient
+import com.cloudogu.gitops.infrastructure.deployment.HelmStrategy
+
+import org.junit.jupiter.api.Test
+import retrofit2.Call
+import retrofit2.Response
 
 class ScmManagerSetupTest {
